Fix multiple sphinx errors; move problem to folder
