+++ pypesto/problem/__init__.py
@@ -1,0 +1,10 @@
+"""
+Problem
+=======
+
+A problem contains the objective as well as all information like prior
+describing the problem to be solved.
+
+"""
+
+from .base import Problem
